+++ src/single.rs
@@ -1,7 +1,26 @@
 //! An accessor to a single element
 
-use crate::{error::Error, mapper::Mapper};
-use core::{fmt, hash::Hash, marker::PhantomData, mem, ptr};
+use {
+    crate::{
+        error::Error,
+        mapper::Mapper,
+        marker::{self, AccessorTypeSpecifier, Readable, Writable},
+    },
+    core::{fmt, hash::Hash, marker::PhantomData, mem, ptr},
+};
+
+/// An alias of [`ReadWrite`].
+#[deprecated(since = "0.3.2", note = "Use `ReadWrite`.")]
+pub type Single<T, M> = ReadWrite<T, M>;
+
+/// A readable and writable accessor.
+pub type ReadWrite<T, M> = Generic<T, M, marker::ReadWrite>;
+
+/// A read-only accessor.
+pub type ReadOnly<T, M> = Generic<T, M, marker::ReadOnly>;
+
+/// A write-only accessor.
+pub type WriteOnly<T, M> = Generic<T, M, marker::WriteOnly>;
 
 /// An accessor to read, modify, and write a single value of memory.
 ///
@@ -9,6 +28,7 @@ use core::{fmt, hash::Hash, marker::PhantomData, mem, ptr};
 ///
 /// ```no_run
 /// use accessor::mapper::Mapper;
+/// use accessor::single;
 /// use core::num::NonZeroUsize;
 ///
 /// struct M;
@@ -25,7 +45,7 @@ use core::{fmt, hash::Hash, marker::PhantomData, mem, ptr};
 /// let mapper = M;
 ///
 /// // Create an accessor to the i32 value at the physical address 0x1000.
-/// let mut a = unsafe { accessor::Single::<i32, M>::new(0x1000, mapper) };
+/// let mut a = unsafe { single::ReadWrite::<i32, M>::new(0x1000, mapper) };
 ///
 /// // Read a value.
 /// a.read_volatile();
@@ -38,20 +58,23 @@ use core::{fmt, hash::Hash, marker::PhantomData, mem, ptr};
 ///     *v *= 2;
 /// });
 /// ```
-pub struct Single<T, M>
+pub struct Generic<T, M, A>
 where
     T: Copy,
     M: Mapper,
+    A: AccessorTypeSpecifier,
 {
     virt: usize,
     bytes: usize,
     _marker: PhantomData<T>,
+    _readable_writable: PhantomData<A>,
     mapper: M,
 }
-impl<T, M> Single<T, M>
+impl<T, M, A> Generic<T, M, A>
 where
     T: Copy,
     M: Mapper,
+    A: AccessorTypeSpecifier,
 {
     /// Creates a new accessor to an element of type `T` at the physical address `phys_base`.
     ///
@@ -75,6 +98,7 @@ where
             virt,
             bytes,
             _marker: PhantomData,
+            _readable_writable: PhantomData,
             mapper,
         }
     }
@@ -102,19 +126,31 @@ where
             })
         }
     }
-
+}
+impl<T, M, A> Generic<T, M, A>
+where
+    T: Copy,
+    M: Mapper,
+    A: Readable,
+{
     /// Reads a value from the address that the accessor points to.
     pub fn read_volatile(&self) -> T {
         // SAFETY: `Accessor::new` ensures that `self.virt` is aligned properly.
         unsafe { ptr::read_volatile(self.virt as *const _) }
     }
 
-    /// Alias of [`Single::read_volatile`].
+    /// Alias of [`Generic::read_volatile`].
     #[deprecated(since = "0.3.1", note = "use `read_volatile`")]
     pub fn read(&self) -> T {
         self.read_volatile()
     }
-
+}
+impl<T, M, A> Generic<T, M, A>
+where
+    T: Copy,
+    M: Mapper,
+    A: Writable,
+{
     /// Writes a value to the address that the accessor points to.
     pub fn write_volatile(&mut self, v: T) {
         // SAFETY: `Accessor::new` ensures that `self.virt` is aligned properly.
@@ -123,12 +159,18 @@ where
         }
     }
 
-    /// Alias of [`Single::write_volatile`].
+    /// Alias of [`Generic::write_volatile`].
     #[deprecated(since = "0.3.1", note = "use `write_volatile`")]
     pub fn write(&mut self, v: T) {
         self.write_volatile(v);
     }
-
+}
+impl<T, M, A> Generic<T, M, A>
+where
+    T: Copy,
+    M: Mapper,
+    A: Readable + Writable,
+{
     /// Updates a value that the accessor points by reading it, modifying it, and writing it.
     ///
     /// Note that some MMIO regions (e.g. the Command Ring Pointer field of the Command
@@ -143,7 +185,7 @@ where
         self.write_volatile(v);
     }
 
-    /// Alias of [`Single::update_volatile`].
+    /// Alias of [`Generic::update_volatile`].
     #[deprecated(since = "0.3.1", note = "use `update_volatile`")]
     pub fn update<U>(&mut self, f: U)
     where
@@ -152,61 +194,68 @@ where
         self.update_volatile(f);
     }
 }
-impl<T, M> fmt::Debug for Single<T, M>
+impl<T, M, A> fmt::Debug for Generic<T, M, A>
 where
     T: Copy + fmt::Debug,
     M: Mapper,
+    A: Readable,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{:?}", self.read_volatile())
     }
 }
-impl<T, M> PartialEq for Single<T, M>
+impl<T, M, A> PartialEq for Generic<T, M, A>
 where
     T: Copy + PartialEq,
     M: Mapper,
+    A: Readable,
 {
     fn eq(&self, other: &Self) -> bool {
         self.read_volatile().eq(&other.read_volatile())
     }
 }
-impl<T, M> Eq for Single<T, M>
+impl<T, M, A> Eq for Generic<T, M, A>
 where
     T: Copy + Eq,
     M: Mapper,
+    A: Readable,
 {
 }
-impl<T, M> PartialOrd for Single<T, M>
+impl<T, M, A> PartialOrd for Generic<T, M, A>
 where
     T: Copy + PartialOrd,
     M: Mapper,
+    A: Readable,
 {
     fn partial_cmp(&self, other: &Self) -> Option<core::cmp::Ordering> {
         self.read_volatile().partial_cmp(&other.read_volatile())
     }
 }
-impl<T, M> Ord for Single<T, M>
+impl<T, M, A> Ord for Generic<T, M, A>
 where
     T: Copy + Ord,
     M: Mapper,
+    A: Readable,
 {
     fn cmp(&self, other: &Self) -> core::cmp::Ordering {
         self.read_volatile().cmp(&other.read_volatile())
     }
 }
-impl<T, M> Hash for Single<T, M>
+impl<T, M, A> Hash for Generic<T, M, A>
 where
     T: Copy + Hash,
     M: Mapper,
+    A: Readable,
 {
     fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
         self.read_volatile().hash(state);
     }
 }
-impl<T, M> Drop for Single<T, M>
+impl<T, M, A> Drop for Generic<T, M, A>
 where
     T: Copy,
     M: Mapper,
+    A: AccessorTypeSpecifier,
 {
     fn drop(&mut self) {
         self.mapper.unmap(self.virt, self.bytes);